// src/generated/graphql.tsx
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** Parse string date to type Date */
  CacheDate: any;
  /** The javascript `Date` as string. Type represents date and time as the ISO Date string. */
  DateTime: any;
  /** The `JSONObject` scalar type represents JSON objects as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSONObject: any;
};
// ...
export type Challenge = {
  __typename?: 'Challenge';
  id: Scalars['ID'];
  name: Scalars['String'];
  challenge_type: Scalars['String'];
  start_date: Scalars['CacheDate'];
  end_date: Scalars['CacheDate'];
  start_date_registration: Scalars['CacheDate'];
  end_date_registration: Scalars['CacheDate'];
  has_achievement: Scalars['Boolean'];
  internal_test: Scalars['Boolean'];
  eligible_spotlight?: Maybe<Scalars['Boolean']>;
  description: Scalars['String'];
  image_avatar: Scalars['String'];
  image_spotlight?: Maybe<Scalars['String']>;
  image_cover: Scalars['String'];
  rank: Scalars['Boolean'];
  count_subscribe: Scalars['Float'];
  accept_installments: Scalars['Boolean'];
  count_unsubscribed: Scalars['Float'];
  count_view: Scalars['Float'];
  claps_count: Scalars['Float'];
  count_comments: Scalars['Float'];
  commissioning_model_id?: Maybe<Scalars['String']>;
  commissioning_model?: Maybe<CommissioningModelType>;
  products?: Maybe<Array<ProductType>>;
  claps?: Maybe<Array<Clap>>;
  comments?: Maybe<Array<Comment>>;
  company_id?: Maybe<Scalars['String']>;
  creator_id?: Maybe<Scalars['String']>;
  prestart_visibility: Scalars['Boolean'];
  isFavorite?: Maybe<Scalars['Boolean']>;
  summary: ChallengeSummary;
  launch_date?: Maybe<Scalars['CacheDate']>;
  registration_limit?: Maybe<Scalars['Int']>;
  date_for_drawn?: Maybe<Scalars['CacheDate']>;
  date_of_drawn?: Maybe<Scalars['CacheDate']>;
  physical_event: Scalars['Boolean'];
  has_categories?: Maybe<Scalars['Boolean']>;
  temporarily_unavailable: Scalars['Boolean'];
  published: Scalars['Boolean'];
  user_challenges: Array<UserChallenges>;
  winners?: Maybe<Array<UserChallenges>>;
  awards?: Maybe<Array<ChallengeAwards>>;
  configuration?: Maybe<ChallengeConfiguration>;
  company: Company;
  products_purchased_without_subscription: Array<ProductPurchasedType>;
  subscription_prices: Array<EventSubscriptionPriceType>;
  creator: User;
  challenge_images: Array<ChallengeImages>;
  challenges_attached_files: Array<ChallengeAttachedFiles>;
  challenges_external_links_attached: Array<ChallengeExternalLinksAttached>;
  challenge_categories: Array<ChallengeCategories>;
  user_extraordinary_actions?: Maybe<Array<UserEventExtraordinaryActionType>>;
  support_points?: Maybe<Array<EventSupportPointType>>;
  event_routes?: Maybe<Array<EventRoutesType>>;
};
// ...
export type City = {
  __typename?: 'City';
  id: Scalars['ID'];
  name: Scalars['String'];
  slug: Scalars['String'];
  status: Scalars['Boolean'];
  id_locale?: Maybe<Scalars['String']>;
  state: State;
};
// ...
export type LoginResponse = {
  __typename?: 'LoginResponse';
  user: User;
  profile?: Maybe<Profile>;
  accessToken: Scalars['String'];
  refreshToken: Scalars['String'];
};

export type LoginUserInput = {
  email: Scalars['String'];
  password: Scalars['String'];
};

export type MarkSubscriptionAsWithdrawnInput = {
  challenge_id: Scalars['String'];
  subscription_user_id: Scalars['String'];
  user_id_who_marked_the_withdraw: Scalars['String'];
};

export type Mutation = {
// ...
  login: LoginResponse;
// ...
};
// ...
export type MutationLoginArgs = {
  data: LoginUserInput;
};
// ...
export type Payment = {
  __typename?: 'Payment';
  id: Scalars['String'];
  resource_payment_id: Scalars['String'];
  origin_payment_id: Scalars['String'];
  payment_processor_id: Scalars['String'];
  country_id: Scalars['String'];
  bill_expiration_date?: Maybe<Scalars['CacheDate']>;
  value: Scalars['Float'];
  is_paid: Scalars['Boolean'];
  declined: Scalars['Boolean'];
  returned: Scalars['Boolean'];
  entrance: Scalars['Boolean'];
  processed: Scalars['Boolean'];
  origin_resource_id?: Maybe<Scalars['String']>;
  user_card_id?: Maybe<Scalars['String']>;
  profile_id: Scalars['String'];
  user_id: Scalars['String'];
  user: User;
  bill_link?: Maybe<Scalars['String']>;
  status: Scalars['String'];
  refuse_reason?: Maybe<Scalars['String']>;
  humanized_message?: Maybe<Scalars['String']>;
  acquirer_return_code: Scalars['String'];
  acquirer_name?: Maybe<Scalars['String']>;
  tid: Scalars['String'];
  nsu: Scalars['String'];
  cost?: Maybe<Scalars['String']>;
  antifraud_score?: Maybe<Scalars['String']>;
  bill_barcode?: Maybe<Scalars['String']>;
  authorization_code?: Maybe<Scalars['String']>;
  authorized_amount?: Maybe<Scalars['Float']>;
  paid_amount?: Maybe<Scalars['Float']>;
  refunded_amount?: Maybe<Scalars['Float']>;
  installments?: Maybe<Scalars['Float']>;
  split_rules?: Maybe<Scalars['String']>;
  antifraud_metadata?: Maybe<Scalars['String']>;
  antifraud_order_id?: Maybe<Scalars['String']>;
  pix_qrcode?: Maybe<Scalars['String']>;
  pix_expiration_date?: Maybe<Scalars['CacheDate']>;
  effective_payment_date?: Maybe<Scalars['CacheDate']>;
  interest_free_amount?: Maybe<Scalars['Float']>;
  created_at: Scalars['CacheDate'];
  updated_at: Scalars['CacheDate'];
  origin_payment: Origin;
  card?: Maybe<Card>;
  product_purchased_user_payment?: Maybe<ProductPurchasedUserPaymentType>;
};
// ...
export type Profile = {
  __typename?: 'Profile';
  id: Scalars['ID'];
  username: Scalars['String'];
  followers_count: Scalars['Float'];
  following_count: Scalars['Float'];
  official: Scalars['Boolean'];
  profile_avatar?: Maybe<Scalars['String']>;
  profile_cover?: Maybe<Scalars['String']>;
  description?: Maybe<Scalars['String']>;
  user_id: Scalars['String'];
  company_id: Scalars['String'];
  is_follower?: Maybe<Scalars['Boolean']>;
  user: User;
};

export type ProfileSchedule = {
  __typename?: 'ProfileSchedule';
  _id: Scalars['ID'];
  profile_id: Scalars['String'];
  resource_key: Scalars['String'];
  date_occurrence?: Maybe<Scalars['DateTime']>;
  title: Scalars['String'];
  subtitle: Scalars['String'];
  deep_link: Scalars['String'];
  web_link: Scalars['String'];
  image_link: Scalars['String'];
};

export type Query = {
// ...
  checkEmail: User;
// ...
};
// ...
export type QueryCheckEmailArgs = {
  email: Scalars['String'];
};
// ...
export type User = {
  __typename?: 'User';
  id: Scalars['ID'];
  name?: Maybe<Scalars['String']>;
  firstname?: Maybe<Scalars['String']>;
  lastname?: Maybe<Scalars['String']>;
  email?: Maybe<Scalars['String']>;
  gender?: Maybe<Scalars['String']>;
  stature?: Maybe<Scalars['Float']>;
  weight?: Maybe<Scalars['Float']>;
  phone?: Maybe<Scalars['String']>;
  date_of_birth?: Maybe<Scalars['CacheDate']>;
  has_social_login: Scalars['Boolean'];
  city_id?: Maybe<Scalars['String']>;
  city?: Maybe<City>;
  active: Scalars['Boolean'];
  address_line_one?: Maybe<Scalars['String']>;
  address_line_two?: Maybe<Scalars['String']>;
  street_number?: Maybe<Scalars['String']>;
  legal_registry_number?: Maybe<Scalars['String']>;
  zip_code?: Maybe<Scalars['String']>;
  strava_permission_activities: Scalars['Boolean'];
  staff: Scalars['Boolean'];
  blacklist: Scalars['Boolean'];
  team_name?: Maybe<Scalars['String']>;
  companies: Array<UsersCompanies>;
  payments: Array<Payment>;
  created_at: Scalars['DateTime'];
  updated_at: Scalars['DateTime'];
  profile?: Maybe<Profile>;
  last_challenge?: Maybe<Challenge>;
  count_challenges_participates?: Maybe<Scalars['Float']>;
  user_event_extraordinary_action?: Maybe<Array<UserEventExtraordinaryActionType>>;
  activities_count?: Maybe<Scalars['Float']>;
};
// ...
export type UserEventExtraordinaryActionType = {
  __typename?: 'UserEventExtraordinaryActionType';
  id: Scalars['ID'];
  user_id: Scalars['String'];
  challenge_id: Scalars['String'];
  bonus_subscription?: Maybe<Scalars['Boolean']>;
  buy_after_registration_closes?: Maybe<Scalars['Boolean']>;
  user: User;
  challenge?: Maybe<Challenge>;
  active: Scalars['Boolean'];
  created_at: Scalars['CacheDate'];
  updated_at: Scalars['CacheDate'];
};
// ...
export type UsersCompanies = {
  __typename?: 'UsersCompanies';
  creator: Scalars['Boolean'];
  company_id: Scalars['String'];
  user_id: Scalars['String'];
  company_function_id: Scalars['String'];
  company: Company;
  user: User;
};